+++ awscli/customizations/datapipeline/__init__.py
@@ -7,10 +7,31 @@ from awscli.customizations.commands import BasicCommand
 from awscli.customizations.datapipeline import translator
 
 
-HELP_TEXT = """\
+DEFINITION_HELP_TEXT = """\
 The JSON pipeline definition.  If the pipeline definition
 is in a file you can use the file://<filename> syntax to
 specify a filename.
+"""
+PARAMETER_OBJECTS_HELP_TEXT = """\
+The JSON parameter objects.  If the parameter objects are
+in a file you can use the file://<filename> syntax to
+specify a filename. You can optionally provide these in
+pipeline definition as well. Parameter objects provided
+on command line would replace the one in definition.
+"""
+PARAMETER_VALUES_HELP_TEXT = """\
+The JSON parameter values.  If the parameter values are
+in a file you can use the file://<filename> syntax to
+specify a filename. You can optionally provide these in
+pipeline definition as well. Parameter values provided
+on command line would replace the one in definition.
+"""
+INLINE_PARAMETER_VALUES_HELP_TEXT = """\
+The JSON parameter values. You can specify these as
+key-value pairs in the key=value format. Multiple parameters
+are separated by a space. For list type parameter values
+you can use the same key name and specify each value as
+a key value pair. e.g. arrayValue=value1 arrayValue=value2
 """
 
 
@@ -18,10 +39,20 @@ class DocSectionNotFoundError(Exception):
     pass
 
 
+class ParameterDefinitionError(Exception):
+    def __init__(self, msg):
+        full_msg = ("Error in parameter: %s\n" % msg)
+        super(ParameterDefinitionError, self).__init__(full_msg)
+        self.msg = msg
+
+
 def register_customizations(cli):
     cli.register(
         'building-argument-table.datapipeline.put-pipeline-definition',
         add_pipeline_definition)
+    cli.register(
+        'building-argument-table.datapipeline.activate-pipeline',
+        activate_pipeline_definition)
     cli.register(
         'after-call.datapipeline.GetPipelineDefinition',
         translate_definition)
@@ -63,21 +94,49 @@ def document_translation(help_command, **kwargs):
 
 def add_pipeline_definition(argument_table, **kwargs):
     argument_table['pipeline-definition'] = PipelineDefinitionArgument(
-        'pipeline-definition', required=True, help_text=HELP_TEXT)
+        'pipeline-definition', required=True,
+        help_text=DEFINITION_HELP_TEXT)
+
+    argument_table['parameter-objects'] = ParameterObjectsArgument(
+        'parameter-objects', required=False,
+        help_text=PARAMETER_OBJECTS_HELP_TEXT)
+
+    argument_table['parameter-values-uri'] = ParameterValuesArgument(
+        'parameter-values-uri',
+        required=False,
+        help_text=PARAMETER_VALUES_HELP_TEXT)
+
+    # Need to use an argument model for inline parameters to accept a list
+    argument_table['parameter-values'] = ParameterValuesInlineArgument(
+        'parameter-values',
+        required=False,
+        nargs='+',
+        help_text=INLINE_PARAMETER_VALUES_HELP_TEXT)
+
     # The pipeline-objects is no longer needed required because
     # a user can provide a pipeline-definition instead.
     # get-pipeline-definition also displays the output in the
     # translated format.
+
     del argument_table['pipeline-objects']
 
 
+def activate_pipeline_definition(argument_table, **kwargs):
+    argument_table['parameter-values-uri'] = ParameterValuesArgument(
+        'parameter-values-uri', required=False,
+        help_text=PARAMETER_VALUES_HELP_TEXT)
+
+    # Need to use an argument model for inline parameters to accept a list
+    argument_table['parameter-values'] = ParameterValuesInlineArgument(
+        'parameter-values',
+        required=False,
+        nargs='+',
+        help_text=INLINE_PARAMETER_VALUES_HELP_TEXT,
+        )
+
+
 def translate_definition(parsed, **kwargs):
-    api_objects = parsed.pop('pipelineObjects', None)
-    if api_objects is None:
-        return
-    else:
-        definition = translator.api_to_definition(api_objects)
-        parsed['objects'] = definition['objects']
+    translator.api_to_definition(parsed)
 
 
 def convert_described_objects(api_describe_objects, sort_key_func=None):
@@ -122,10 +181,10 @@ class QueryArgBuilder(object):
             end_time_str = end_datetime.strftime('%Y-%m-%dT%H:%M:%S')
             selectors.append({
                 'fieldName': '@actualStartTime',
-                    'operator': {
-                        'type': 'BETWEEN',
-                        'values': [start_time_str, end_time_str]
-                    }
+                'operator': {
+                    'type': 'BETWEEN',
+                    'values': [start_time_str, end_time_str]
+                }
             })
         else:
             self._build_schedule_times(selectors, parsed_args)
@@ -140,20 +199,20 @@ class QueryArgBuilder(object):
             end_time_str = parsed_args.start_interval[1]
             selectors.append({
                 'fieldName': '@actualStartTime',
-                    'operator': {
-                        'type': 'BETWEEN',
-                        'values': [start_time_str, end_time_str]
-                    }
+                'operator': {
+                    'type': 'BETWEEN',
+                    'values': [start_time_str, end_time_str]
+                }
             })
         if parsed_args.schedule_interval is not None:
             start_time_str = parsed_args.schedule_interval[0]
             end_time_str = parsed_args.schedule_interval[1]
             selectors.append({
                 'fieldName': '@scheduleStartTime',
-                    'operator': {
-                        'type': 'BETWEEN',
-                        'values': [start_time_str, end_time_str]
-                    }
+                'operator': {
+                    'type': 'BETWEEN',
+                    'values': [start_time_str, end_time_str]
+                }
             })
 
     def _build_status(self, selectors, parsed_args):
@@ -165,13 +224,79 @@ class QueryArgBuilder(object):
             }
         })
 
+
 class PipelineDefinitionArgument(CustomArgument):
     def add_to_params(self, parameters, value):
         if value is None:
             return
         parsed = json.loads(value)
-        api_objects = translator.definition_to_api(parsed)
+        api_objects = translator.definition_to_api_objects(parsed)
+        parameter_objects = translator.definition_to_api_parameters(parsed)
+        parameter_values = translator.definition_to_parameter_values(parsed)
         parameters['pipelineObjects'] = api_objects
+        # Use Parameter objects and values from def if not already provided
+        if 'parameterObjects' not in parameters \
+                and parameter_objects is not None:
+            parameters['parameterObjects'] = parameter_objects
+        if 'parameterValues' not in parameters \
+                and parameter_values is not None:
+            parameters['parameterValues'] = parameter_values
+
+
+class ParameterObjectsArgument(CustomArgument):
+    def add_to_params(self, parameters, value):
+        if value is None:
+            return
+        parsed = json.loads(value)
+        parameter_objects = translator.definition_to_api_parameters(parsed)
+        parameters['parameterObjects'] = parameter_objects
+
+
+class ParameterValuesArgument(CustomArgument):
+    def add_to_params(self, parameters, value):
+
+        if value is None:
+            return
+
+        if parameters.get('parameterValues', None) is not None:
+            raise Exception(
+                "Only parameter-values or parameter-values-uri is allowed"
+            )
+
+        parsed = json.loads(value)
+        parameter_values = translator.definition_to_parameter_values(parsed)
+        parameters['parameterValues'] = parameter_values
+
+
+class ParameterValuesInlineArgument(CustomArgument):
+    def add_to_params(self, parameters, value):
+
+        if value is None:
+            return
+
+        if parameters.get('parameterValues', None) is not None:
+            raise Exception(
+                "Only parameter-values or parameter-values-uri is allowed"
+            )
+
+        parameter_object = {}
+        # break string into = point
+        for argument in value:
+            try:
+                argument_components = argument.split('=', 1)
+                key = argument_components[0]
+                value = argument_components[1]
+                if key in parameter_object:
+                    parameter_object[key] = [parameter_object[key], value]
+                else:
+                    parameter_object[key] = value
+            except IndexError:
+                raise ParameterDefinitionError(
+                    "Invalid inline parameter format: %s" % argument
+                )
+        parsed = {'values': parameter_object}
+        parameter_values = translator.definition_to_parameter_values(parsed)
+        parameters['parameterValues'] = parameter_values
 
 
 class ListRunsCommand(BasicCommand):
@@ -182,10 +307,11 @@ class ListRunsCommand(BasicCommand):
         'results to include only the runs you are interested in.')
     ARG_TABLE = [
         {'name': 'pipeline-id', 'help_text': 'The identifier of the pipeline.',
-         'action': 'store', 'required': True, 'cli_type_name': 'string',},
+         'action': 'store', 'required': True, 'cli_type_name': 'string', },
         {'name': 'status',
          'help_text': (
-             'Filters the list to include only runs in the specified statuses. '
+             'Filters the list to include only runs in the '
+             'specified statuses. '
              'The valid statuses are as follows: waiting, pending, cancelled, '
              'running, finished, failed, waiting_for_runner, '
              'and waiting_on_dependencies. You can combine statuses as a '
@@ -196,12 +322,12 @@ class ListRunsCommand(BasicCommand):
          'help_text': (
              'Filters the list to include only runs that started '
              'within the specified interval.'),
-         'action': 'store', 'required': False, 'cli_type_name': 'string',},
+         'action': 'store', 'required': False, 'cli_type_name': 'string', },
         {'name': 'schedule-interval',
          'help_text': (
              'Filters the list to include only runs that are scheduled to '
              'start within the specified interval.'),
-         'action': 'store', 'required': False, 'cli_type_name': 'string',},
+         'action': 'store', 'required': False, 'cli_type_name': 'string', },
     ]
     VALID_STATUS = ['waiting', 'pending', 'cancelled', 'running',
                     'finished', 'failed', 'waiting_for_runner',
@@ -233,13 +359,16 @@ class ListRunsCommand(BasicCommand):
         # Parse the status csv.
         if parsed_args.start_interval is not None:
             parsed_args.start_interval = [
-                arg.strip() for arg in parsed_args.start_interval.split(',')]
+                arg.strip() for arg in
+                parsed_args.start_interval.split(',')]
         if parsed_args.schedule_interval is not None:
             parsed_args.schedule_interval = [
-                arg.strip() for arg in parsed_args.schedule_interval.split(',')]
+                arg.strip() for arg in
+                parsed_args.schedule_interval.split(',')]
         if parsed_args.status is not None:
             parsed_args.status = [
-                arg.strip() for arg in parsed_args.status.split(',')]
+                arg.strip() for arg in
+                parsed_args.status.split(',')]
             self._validate_status_choices(parsed_args.status)
 
     def _validate_status_choices(self, statuses):